requirements.txt
Django==2.2
Djangorestframework==3.10.0
django-cors-headers